Add files via upload
Histogram of expressed opinions when 40% of users with positive attitudes left the social network. The peak in the graph represents the consensus of a group of commenting users around an attitude. The Blue bar charts depict the number of commenting users when 500 users were added to the network, while the red charts represent the number of commenting users at the final state with 1000 users.