Add files via upload
Snapshots of the network in a state where only normal users are present. Without the presence of noisy users, the network is usually without polarity and will reach a consensus after the passage of time. Of course, it should be noted that consensus means the consensus of users who express their opinion. For this purpose, during visualization, users whose expression-status parameter is 1 should be filtered.